Assignment-2/REST_API/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
mysqlclient==2.1.0
python-decouple==3.5
